Replase masquerade cntroller to users module
