+++ app/views/admin/users/index.html.slim
@@ -24,9 +24,9 @@
               td
                 = user.email
               td
-                = link_to("Update", edit_admin_user_path(id: user.id))
+                = link_to("Update", edit_admin_user_path(user))
               td
-                = link_to(user.new_session, new_admin_masquerade_path(user_id: user.id))
+                = link_to(user.new_session, new_admin_user_masquerade_path(user))
       = paginate users
   .medium-3.columns data-sticky-container=""
     = render "tools_menu"
