+++ app/views/application/_navigation_main.html.slim
@@ -4,6 +4,6 @@ ul.left
     = active_link_to "Subscriptions", user_subscriptions_path(current_user),  wrap_tag: :li
   = active_link_to "Results", results_path,  wrap_tag: :li
   = active_link_to "Hashtags", hashtags_path,  wrap_tag: :li
-  - if masquerading?
-    = active_link_to "Stop Masquerading", admin_masquerade_path(current_user.id), method: :delete, wrap_tag: :li
+  - if masquerading? && user_signed_in?
+    = active_link_to "Stop Masquerading", admin_user_masquerade_path(current_user, current_user), method: :delete, wrap_tag: :li
